Remove and ignore files generated by autotools.
Leaving auto-generated files sometimes cause problems with newer
changes.

In released packages (tarballs) these file is necessary for users to
build without autotools. We don't develop softwares in the directory
that is extracted from a tarball so the problems won't happen.

However, we, developers living in the Linux world, already have
installed autotools, and the working tree from a repository is the place
where we develop softwares.

We might have to expand the installation instructions but I just
appended the `autoreconf -i` line for now.

Signed-off-by: OGINO Masanori <masanori.ogino@gmail.com>