Removed other variables related to displaying IK debugging info
To display the location of IK vertices (or any vertices) in general, you have to create VAO and buffer objects.
The messy IK debugging code I wrote had all the variables for creating the VAO/Buffer objects for the IK debug display in the Viewer class.
As we have removed displayIKDebug() (or whatever it was called), we should remove these VAO/Buffer objects as well.